+++ aws-android-sdk-auth-core/src/main/java/com/amazonaws/mobile/auth/core/signin/SignInManager.java
@@ -20,7 +20,6 @@ package com.amazonaws.mobile.auth.core.signin;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.util.SparseArray;
 import android.view.View;